requirements.txt
certifi==2018.10.15
chardet==3.0.4
idna==2.7
python-dotenv==0.9.1
requests==2.20.0
tornado==5.1.1
urllib3==1.24.1
